# swagger/story_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "story_service.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/89minutes/create": {
      "post": {
        "operationId": "StoryService_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/CreateStoryResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateStoryRequest"
            }
          }
        ],
        "tags": [
          "StoryService"
        ]
      }
    },
    "/v1/89minutes/ping": {
      "post": {
        "operationId": "StoryService_PingPong",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/Pong"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Ping"
            }
          }
        ],
        "tags": [
          "StoryService"
        ]
      }
    }
  },
  "definitions": {
    "CreateStoryRequest": {
      "type": "object",
      "properties": {
        "story": {
          "$ref": "#/definitions/Story"
        }
      }
    },
    "CreateStoryResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        }
      }
    },
    "GetLatestStoryResponse": {
      "type": "object",
      "properties": {
        "story": {
          "$ref": "#/definitions/Story"
        }
      }
    },
    "Ping": {
      "type": "object",
      "properties": {
        "ping": {
          "type": "string"
        }
      }
    },
    "Pong": {
      "type": "object",
      "properties": {
        "ping": {
          "type": "string"
        }
      }
    },
    "Story": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "UUID"
        },
        "title": {
          "type": "string",
          "title": "Title of the story"
        },
        "textContent": {
          "type": "string",
          "title": "Text contents"
        },
        "fileContent": {
          "type": "array",
          "items": {
            "type": "string",
            "format": "byte"
          },
          "title": "Story can have multiple images and videos"
        },
        "formattedText": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Story can have formatted texts, like code"
        },
        "WrittenBy": {
          "type": "string",
          "title": "Name of the author"
        },
        "isDraft": {
          "type": "boolean",
          "title": "Is the story draft and not published"
        },
        "Tag": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Tag"
          },
          "title": "Tags for categorizing stories"
        },
        "createTime": {
          "type": "string",
          "format": "date-time",
          "title": "Create time"
        },
        "updateTime": {
          "type": "string",
          "format": "date-time",
          "title": "Update time"
        }
      }
    },
    "Tag": {
      "type": "object",
      "properties": {
        "tagName": {
          "type": "string"
        }
      }
    },
    "UpdateStoryResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "UUID"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "runtimeStreamError": {
      "type": "object",
      "properties": {
        "grpc_code": {
          "type": "integer",
          "format": "int32"
        },
        "http_code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "http_status": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}